Update Api.php
This code breaks with HHVM because current() does not behave the same way in PHP. The Mage_Core_Model_Config_Element Object is essentially a SimpleXML object, which is traversable, but not an Array.  The current() function states it works on arrays, not traversable objects.  To address this we can cast the node to a string, which is a common practice throughout magento's codebase.